requirements.txt
# Dependencias del proyecto TCG Web App

# Backend
fastapi==0.128.0
uvicorn==0.40.0
supabase==2.27.0
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1

# Scraper
requests==2.32.5
beautifulsoup4==4.13.4
lxml==6.0.2
aiohttp==3.11.0
pandas==2.3.3
numpy==2.4.0

# Testing
pytest==8.4.1
pytest-mock==3.14.1

# Utilities
colorama==0.4.6
pygments==2.19.2
rich==14.2.0
python-dateutil==2.9.0
